requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document Processing
unstructured[all-docs]==0.11.6
python-docx==1.1.0
PyPDF2==3.0.1
pymupdf==1.23.8
python-magic-bin==0.4.14  # Windows support
filetype==1.2.0

# Embeddings and Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# LLM and NLP
transformers==4.36.0
torch==2.1.0
accelerate==0.24.1
bitsandbytes==0.41.3
optimum==1.15.0
tokenizers==0.15.0

# Text Processing and Utilities
langchain==0.0.350
langchain-community==0.0.1
tiktoken==0.5.2
nltk==3.8.1
spacy==3.7.2

# HTTP Client and Async
httpx==0.25.2
aiofiles==23.2.1
asyncio==3.4.3

# Groq API
groq==0.4.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
